[archive] add link follow-up to FileCacheArchive.add_link()
Creating a link may trigger further actions in the archive: if the
link target is a regular file, we must copy that file into the
archive, and if the target is a symbolic link, then we must create
that link, and copy in the link target.

Handle this by calling add_file() or (recursively) add_link() in
order to create the missing pieces of the symlink chain.

These operations must take place outside of the path lock since
they do not modify the archive namespace and will call methods of
the Archive object that will attempt to re-acquire this lock.

Signed-off-by: Bryn M. Reeves <bmr@redhat.com>